add hp to player info
